requirements.txt
pandas
geopandas
shapely
pyyaml
scikit-learn
xgboost
joblib
numpy
fastapi
uvicorn
pyarrow
matplotlib